ELEMENTS-1311: refactor code panel template for storybook stories
